% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_bad_argument}
\alias{abort_bad_argument}
\title{Consistent and informative error}
\usage{
abort_bad_argument(arg, must, not = NULL)
}
\arguments{
\item{arg}{The name of the argument not being correct}

\item{must}{Description of what it should be}

\item{not}{Description of what it should NOT be

Taken from https://adv-r.hadley.nz/conditions.html#signalling-conditions}
}
\description{
Consistent and informative error
}
