% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Automatically annotate a raw single animal behavioral dataset}
\usage{
annotate(out, FUN = NULL, updateProgress = NULL)
}
\arguments{
\item{out}{A raw i.e. one row per frame behavr table}

\item{FUN}{A function or a list of functions describing how to aggregate the data into
non overlapping and consecutive time_windows of a fixed length. It should generate at least one
new column where each value is a summary statistic for the corresponding time window}

\item{updateProgress}{A function to update a shiny::Progress instance}
}
\value{
A behavr table with all summary statistics in different columns
}
\description{
Given a raw behavr table and a list of annotating functions
return a list of behavr tables with one resulting table per input function
}
\details{
All functions should summarise using the same windows
out should contain a single id. If you need to analyze data for several flies
call \code{annotate} like this: \code{data[, annotate(.SD, FUN, progress), by = "id"]}
}
