% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_all}
\alias{annotate_all}
\title{Annotate a behavr table, possibly with more than one id (animal)}
\usage{
annotate_all(data, ...)
}
\arguments{
\item{data}{A multi animal behavr table to be pre analyzed or annotated}

\item{...}{Additional arguments to annotate_single_roi}
}
\description{
Annotate a behavr table, possibly with more than one id (animal)
}
