% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate-single-roi.R
\name{annotate_one_func}
\alias{annotate_one_func}
\title{Annotate data from a single animal using an annotation function}
\usage{
annotate_one_func(data, FUN, ...)
}
\arguments{
\item{data}{A raw i.e. one row per frame behavr table}

\item{FUN}{A function describing how to aggregate the data into
non overlapping and consecutive time_windows of a fixed length, defined on their time_window_length arg.
It must:
\begin{itemize}
\item Generate at least one new column where each value is a summary statistic for the corresponding time window
\end{itemize}}

\item{...}{Additional arguments to FUN}
}
\description{
Annotate data from a single animal using an annotation function
}
