% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_single_roi}
\alias{annotate_single_roi}
\title{Automatically annotate a raw single animal behavioral dataset}
\usage{
annotate_single_roi(
  data,
  FUN = NULL,
  updateProgress = NULL,
  path = NULL,
  region_id = NULL,
  ...
)
}
\arguments{
\item{data}{A raw i.e. one row per frame behavr table}

\item{FUN}{A function or list of functions}

\item{updateProgress}{if passed, this function runs everytime an annotation is completed}

\item{path}{if passed, the function reports the path that the animal was loaded from (which database)}

\item{region_id}{if passed, the function repors the roi number that the animal was loaded from (which position)}

\item{...}{Additional arguments to the passed FUN}
}
\value{
A behavr table with all summary statistics in different columns
}
\description{
Given a raw behavr table and a list of annotating functions
return a list of tables with one resulting table per input function
}
\details{
All functions should summarise using the same windows
out should contain a single id. If you need to analyze data for several flies
call \code{annotate_single_roi} like this: \code{data[, annotate(.SD, FUN, progress), by = "id"]}
or call \code{scopr::annotate_all}. This is needed because annotate_single_roi
is NOT id-aware (because it expects a single animal). As such, the result is NOT a behavr table!
}
\seealso{
annotate_one_func
}
