% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-ethoscope-query.R
\name{buildEthoscopeQuery}
\alias{buildEthoscopeQuery}
\title{Build a functional query for loading ethoscope data using the date of experiments and device names
to retreive result files}
\usage{
buildEthoscopeQuery(result_dir, query = NULL, use_cached = FALSE,
  index_file = NULL)
}
\arguments{
\item{result_dir}{The location of the result directory (i.e. the folder containing all the data).}

\item{query}{An optional query formatted as a dataframe (see details).}

\item{use_cached}{whether cache files should be used}

\item{index_file}{An optional file listing all experiments.}
}
\value{
Query extended with the requested paths.
When \code{query} is not specified, the function returns a table with all available files.
}
\description{
This function is designed to list and select experimental files.
In general, end-users will want to retrieve  path to their experimental files
according to the date and ID of the ethoscope without having to understand the underlying directory structure.
}
\details{
The optional argument \code{query} is expected to be a table where each row maps an experiment.
In many respects, it is similar to the \code{what} argument in \link{loadEthoscopeData}.
The only difference is that it does not have a \code{"path"} column.
Instead, it must contain two columns:
\itemize{
\item \code{date}  the date and time when the experiment started formatted either as \code{yyyy-mm-dd} or
\code{yyyy-mm-dd_hh:mm:ss}. In the former case, there may be several matching experiments to a
single time (starting the same day). When this happens, \strong{only the last} is returned,
and a warning message is displayed.
\item \code{machine_name} The name of the machine that acquired the data.
}

The result is meant to be used directly, as the \code{what} argument, in \link{loadEthoscopeData} (see examples).
}
\note{
The ethoscope platform store data in a hard-coded directory structure
\code{"/root_dir/machine_id/machine_name/datetime/file.db"}, where:
\itemize{
\item \code{machine_id} Is, in principle, a universally unique identifier of the acquisition device.
\item \code{machine_name} a human friendly name for acquisition device. In practice, this is expected to be unique within laboratory.
\item \code{datetime} the date and time of the start of the experiment
}
}
\seealso{
\itemize{
\item \link{loadEthoscopeData}, example case #3
\item Tutorial for this function \url{http://gilestrolab.github.io/rethomics/tutorial/todo}
\item What queries are \url{http://gilestrolab.github.io/rethomics/tutorial/todo}
}
}
