% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-single-roi.R
\name{database_is_available}
\alias{database_is_available}
\title{Can I read the sqlite3 database, or is it locked}
\usage{
database_is_available(FILE)
}
\arguments{
\item{FILE}{Path to sqlite3 database file}
}
\value{
Boolean stating whether the METADATA table can be read or not
}
\description{
Can I read the sqlite3 database, or is it locked
}
