% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-info.R
\name{experiment_info}
\alias{experiment_info}
\title{Retrieve information about an experiment}
\usage{
experiment_info(FILE)
}
\arguments{
\item{FILE}{the name of the input file}
}
\value{
a list containing fields for metadata entries
}
\description{
This function is used to obtain experimental information -- such as \emph{time and date} of the experiment,
\emph{acquisition device}, and \emph{version of the software} --
embedded in a result file (\code{.db}) generated by the ethoscope platform.
}
\seealso{
\itemize{
\item \link{load_ethoscope} -- to load the actual data
\item \link{list_result_files} -- to list available files
}
}
