% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sqlite.R
\name{get_metadata}
\alias{get_metadata}
\title{Load the METADATA table in an ethoscope sqlite3 file}
\usage{
get_metadata(path)
}
\arguments{
\item{path}{Path to the dbfile}
}
\description{
Load the METADATA table in an ethoscope sqlite3 file
}
