% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-single-roi.R
\name{get_ms_after_ref}
\alias{get_ms_after_ref}
\title{Adjust t so t0 is at ZT0}
\usage{
get_ms_after_ref(experiment_info, reference_hour)
}
\arguments{
\item{experiment_info}{list with date_time field in POSIXct format}

\item{reference_hour}{integer, hour of the day when ZT0 occurs in GMT timezone}
}
\value{
modified behavr where the column t has its 0 at the ZT0 of the first day
}
\description{
Adjust t so t0 is at ZT0
}
