% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sqlite.R
\name{get_selected_options}
\alias{get_selected_options}
\title{Parse the json data extracted from the METADATA - selected_options of a dbfile
into an R list}
\usage{
get_selected_options(metadata)
}
\arguments{
\item{metadata}{list produced by reading into R the result of the SELECT * FROM METADATA statement
and putting each row into an element of the list, where the field is the list element's name and the value is the list element's value}
}
\description{
Parse the json data extracted from the METADATA - selected_options of a dbfile
into an R list
}
