% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_writable_path}
\alias{get_writable_path}
\title{Update the path so it is writable for sure}
\usage{
get_writable_path(path)
}
\arguments{
\item{path}{character for a path in the filesystem}
}
\description{
If path is not writable, use the same filename but saved instead
to the $HOME folder, which should always be writable
}
