% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ethoscope.R
\name{load_data}
\alias{load_data}
\title{Load and preanalyze all entries from one dbfile
Several cores can be used (if ncores>1)}
\usage{
load_data(q_l, ncores = 1, ...)
}
\arguments{
\item{ncores}{number of cores to use for optional parallel processing (experimental).}

\item{...}{Extra arguments to be passed to load_row}
}
\value{
behavr table
}
\description{
Load and preanalyze all entries from one dbfile
Several cores can be used (if ncores>1)
}
\seealso{
load_row
}
