% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sqlite.R
\name{load_ethoscope_multiplex}
\alias{load_ethoscope_multiplex}
\title{Load ethoscope data with a set of annotation parameters for different intervals}
\usage{
load_ethoscope_multiplex(..., intervals = NULL)
}
\arguments{
\item{...}{Arguments to load_ethoscope}

\item{intervals}{Named list where every element is the start and end time in seconds since experiment start of an interval
The name of the interval should be interval_X, where X is the interval identifier}
}
\description{
Load ethoscope data with a set of annotation parameters for different intervals
}
\seealso{
load_ethoscope
}
