% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-ethoscope.R
\name{load_row}
\alias{load_row}
\title{Call parse_single_roi with sensible arguments
Prepare a list of arguments to pass to parse_single_roi
using arguments explicitly declared by the user or taken from the metadata}
\usage{
load_row(
  row,
  min_time = 0,
  max_time = Inf,
  verbose = FALSE,
  cache = NULL,
  columns,
  FUN,
  ...
)
}
\arguments{
\item{min_time, }{max_time load only data between \code{min_time} and \code{max_time} (in seconds).
This time is \emph{relative to the start of the experiment}.}

\item{verbose}{whether to print progress (a logical)}

\item{cache}{the name of a local directory to cache results for faster subsequent data loading.}

\item{columns}{optional vector of columns to be selected from the db file.
Time (t) is always implicitly selected.
When \code{NULL} and if \code{FUN} is set, columns can be retrieved automatically (from the attributes of \code{FUN}).}

\item{reference_hour}{hour, in the day, to use as ZT0 reference.
When unspecified, time will be relative to the start of the experiment.}
}
\description{
Call parse_single_roi with sensible arguments
Prepare a list of arguments to pass to parse_single_roi
using arguments explicitly declared by the user or taken from the metadata
}
