% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-index-file.R
\name{make_index_file}
\alias{make_index_file}
\title{Used by platform administrators to make an index file listing all experiments}
\usage{
make_index_file(result_dir, index_file = "index.txt")
}
\arguments{
\item{result_dir}{the root directory where all data are saved}

\item{index_file}{the name of an index_file, in \code{result_dir} (useful for loading remote data).}
}
\value{
the path to the index file written (in \code{result_dir}))
}
\description{
This function builds an index which lists the available experimental file.
It is useful to build indices when outing data on a remote server (e.g. FTP).
}
