% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_annotations.R
\name{merge_annotations}
\alias{merge_annotations}
\title{Merge annotation function results into a single dataframe}
\usage{
merge_annotations(x, y)
}
\description{
This function merges the output of two different annotation functions for the same dataset,
represented by x and y, in such a way that x gains all the columns in y not present in x
The merge is performed using the t column
}
