% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace-filters.R
\name{Mode_filter}
\alias{Mode_filter}
\title{Filter noise in the data by taking the most common value
This is useful to remove outlier data produced due to (sparse) mistracking results
Plese note if mistracking is too frequent, this filter is not useful}
\usage{
Mode_filter(data, columns, ...)
}
\arguments{
\item{data}{A dataframe with some column(s) that need(s) to be filtered}

\item{columns}{Character vector of column names in data.
Error is raised if a passed column name is not available in the data}
}
\value{
Filtered data
}
\description{
Filter noise in the data by taking the most common value
This is useful to remove outlier data produced due to (sparse) mistracking results
Plese note if mistracking is too frequent, this filter is not useful
}
