% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace-filters.R
\name{Mode_filter_one_column}
\alias{Mode_filter_one_column}
\title{Apply the mode filter to one column only}
\usage{
Mode_filter_one_column(data, column, window_width = 10)
}
\arguments{
\item{data}{dataframe}

\item{column}{Name of column to be filtered}

\item{window_width}{Number of datapoints the filter looks at when sweeping the data (~kernel size)}
}
\value{
Filtered data
}
\description{
Apply the mode filter to one column only
}
