% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{no_region_id_query}
\alias{no_region_id_query}
\title{A simple toy query defining three experiments to query, with one condition (\code{test}) per experiment.
Implicitly, 20 animals are in each experiment
It serves as a simple example.}
\format{An object of class \code{data.frame} with 3 rows and 4 columns.}
\usage{
no_region_id_query
}
\description{
A simple toy query defining three experiments to query, with one condition (\code{test}) per experiment.
Implicitly, 20 animals are in each experiment
It serves as a simple example.
}
\author{
Quentin Geissmann
}
\keyword{datasets}
