% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-query.R, R/parse-remote-query.R
\name{parse_query}
\alias{parse_query}
\alias{parse_remote_query}
\title{Build a query for loading ethoscope data}
\usage{
parse_query(x, result_dir = NULL, index_file = NULL)

parse_remote_query(x, remote_dir, result_dir, index_file = "index.txt",
  overwrite_local = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{object such as a \link{data.frame}, or the name of a file (see detail)}

\item{result_dir}{the root directory where all data are saved}

\item{index_file}{the name of an index_file, in \code{result_dir} (useful for loading remote data).}

\item{remote_dir}{the url of the result directory on the data server}

\item{overwrite_local}{whether to download all files.
The default, \code{FALSE}, is to only fetch files if they are newer on the remote.}

\item{verbose}{whether to print progress (a logical)}
}
\value{
a \link[data.table:data.table]{data.table::data.table}.
It contains the columns:
\itemize{
\item \code{path} -- the file to retreive the data from
\item \code{experiment_id} -- an autogenerated unique id for the experiment
\item \code{region_id} -- a unique identifier of each individual (region of interest) within an experiment
\item Variables already present in x (if x was a \code{data.frame}).
}
}
\description{
These functions make a query table, which is essential for subsequent loading of
ethoscope data (via \link{query_ethoscopes}).
}
\details{
These function will generate a query from three different types of inputs:
\enumerate{
\item A vector of names of \code{.db} files (only for local queries with \code{parse_query})
\item A \link{data.frame} (recomended)
In this case, the function will try to match requested data with data available on \code{result_dir}.
The provided \link{data.table} has typically one row per requested individial and the columns
(not necessarily in this order):
\itemize{
\item \code{machine_name} -- the name of the machine in which the individual was (e.g. \code{"ETHOSCOPE_001"})
\item \code{date} -- the start date of the experiment formated as \code{"YYYY-MM-DD"}
\item \code{region_id} -- the ROI in which the animal was. When \emph{not provided, all regions are queried}.
\item \code{time} -- the start time of the experiment formated as "HH:MM:SS".
When \emph{not provided}, and multiple expriment for the same machine exist, \emph{only the last one is loaded}.
\item \code{???} -- any number of arbitrary columns* to associate \code{conditions}/\code{treatments}/\code{genotypes}/... to the previous columns.
}
\item The name of a CSV file that contains a table as described in 2.
}
}
\examples{
# a query with no region_id, all regions will be loaded with the same metadata
dir <- paste0(scopr_example_dir(), "/ethoscope_results/")
data(no_region_id_query)
query <- parse_query(no_region_id_query, dir)
print(query)
# a query with region_id, only stated regions will be loaded with specific metadata
data(region_id_query)
query <- parse_query(region_id_query, dir)
print(query)
}
\references{
\itemize{
\item TODO -- to understand the magical power of the query system!
}
}
\seealso{
\itemize{
\item \link{query_ethoscopes} -- to load the actual data
\item \link{list_result_files} -- to list available file
}
}
