% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-single-roi.R
\name{parse_single_roi_wrapped}
\alias{parse_single_roi_wrapped}
\title{Load data from a sqlite database and preanalyze (annotate) it with FUN}
\usage{
parse_single_roi_wrapped(
  id,
  region_id,
  path,
  min_time = 0,
  max_time = +Inf,
  reference_hour = NULL,
  columns = NULL,
  file_size = 0,
  verbose = FALSE,
  updateProgress_load,
  updateProgress_annotate,
  FUN = NULL,
  ...
)
}
\arguments{
\item{region_id}{numeric, position of the animal's roi in the ethoscope's layout. Should be 1-20 for the standard 20-animal arena}

\item{min_time}{numeric, time in seconds from the start of the experiment after which the data is loaded into R}

\item{max_time}{numeric, time in seconds from the start of the experiment after which the data is not loaded into R}

\item{reference_hour}{integer, hour of the day in GMT TZ when the incubator turned the light on. If the reference hour did not happen yet on the loading day,
a whole extra day is prepended to the data, because the time series always starts with an L phase (even if this L phase is fully NA)}

\item{columns}{Columns to be loaded from the dbfile. Checked against the VAR_MAP for availability}

\item{FUN, }{function or list, a function or list of functions that processes the data loaded from the SQLite file in some meaningful way}
}
\description{
Loading is done by read_single_roi. Annotation is done by annotate_single_roi
}
