% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-ethoscopes.R
\name{query_ethoscopes}
\alias{query_ethoscopes}
\title{Read data from an ethoscope result file}
\usage{
query_ethoscopes(result_dir, query = NULL, min_time = 0, max_time = Inf,
  reference_hour = NULL, verbose = TRUE, columns = NULL, ncores = 1,
  FUN = NULL, ...)
}
\arguments{
\item{result_dir}{the root directory where all daily data are saved}

\item{query}{\link{data.frame} representing a formatted query used to request data (see detail)}

\item{min_time}{Exclude data before min_time (in seconds).
This time is \emph{relative to the start of the experiment}.}

\item{max_time}{Exclude data after max_time (in seconds).
It is also relative to the start of the experiment.}

\item{reference_hour}{Hour, in the day, to use as ZT0 reference.
When unspecified, time will be relative to the start of the experiment.}

\item{verbose}{whether to print progress (a logical).}

\item{columns}{Optionnal vector of columns to be selected from the db file.
Time (t) is always implicitely selected.}

\item{ncores}{Number of cores to use for optionnal parallel processing.}

\item{FUN}{function (optional) to transform the data from each animal
immediately after is has been loaded.}

\item{...}{extra arguments to be passed to \code{FUN}}
}
\value{
A \link{behavr} table.
The metadata contains all the query columns and an autogenerated id per animal.
The data has the columns:
\itemize{
\item \code{id} -- autogenerated unique identifier, one per animal
\item \code{t} -- time
\item Several variables recorded by ethoscopes (position, angle, width/height and others).
Distance units (e.g. xy position, height/width) are expressed as a fraction of the width of the ROI they originate from.
}
}
\description{
This function is used to convert all the information
contained in a result file generated by the \href{http://gilestrolab.github.io/ethoscope/}{ethoscope platform}
(i.e a .db file) into an \code{R} \code{data.table}.
}
\details{
todo
}
\examples{
\donttest{
todo
}
}
\seealso{
TODO
}
