% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Load a metadata .csv into R and perform basic validation}
\usage{
read_metadata(metadata_path)
}
\description{
Load a metadata .csv into R and perform basic validation
}
