% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-single-roi.R
\name{read_single_roi}
\alias{read_single_roi}
\title{Read data from a single ROI in a SQLite file}
\usage{
read_single_roi(
  FILE,
  region_id,
  min_time = 0,
  max_time = +Inf,
  reference_hour = NULL,
  columns = NULL,
  time_stamp = NULL
)
}
\arguments{
\item{FILE}{character, absolute path to a sqlite3 file (ending in .db typically)}

\item{region_id}{numeric, position of the animal's roi in the ethoscope's layout. Should be 1-20 for the standard 20-animal arena}

\item{min_time}{numeric, time in seconds from the start of the experiment after which the data is loaded into R}

\item{max_time}{numeric, time in seconds from the start of the experiment after which the data is not loaded into R}

\item{reference_hour}{integer, hour of the day in GMT TZ when the incubator turned the light on. If the reference hour did not happen yet on the loading day,
a whole extra day is prepended to the data, because the time series always starts with an L phase (even if this L phase is fully NA)}

\item{columns}{Columns to be loaded from the dbfile. Checked against the VAR_MAP for availability}

\item{time_stamp}{Used to memoise the function}
}
\value{

}
\description{
This function performs the following actions
\itemize{
\item{Verify the desired columns are available, by checking the VAR_MAP table, which describes the ROI_X tables.
If the VAR_MAP is corrupted, a warning is emitted and a default VAR_MAP is loaded from scopr internal dataset
}
\item{Verify the desired ROI is available, by checking the ROI_MAP table}
\item{Load the behavioral data for region_id from table ROI_region_id (e.g. ROI_1, ROI_20, ...)}
\item{Adjust the time loaded from the sqlite file if reference_hour is passed. The loaded t is 0 at experiment start. The adjusted time is 0 at reference_hour}
\item{If the start happened > 12 hours since ZT0 (i.e. in the D phase for a standard 12h LD experiment) ..}
\item{If the start happened < 12 hours since ZT0 (i.e. in the L phase for a standard 12h LD experiment) the difference in between is added so t=0 becomes ZT0, and not the experiment start}
\item{Moreover, the unit is switched from ms to s}
\item{The data is filtered, so only data not inferred from previous positions is kept. Alternatively, if the has_interacted column is available and is 1, the data is also kept, even if inferred}
}
}
