% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_id_metadata}
\alias{region_id_metadata}
\title{Simple toy metadata experimental conditions for 15 individuals in three experiments.
the condition (\code{test}) is linked to the experiments, and another condition (\code{treatment}) is per individual.
It serves as a simple example.
#'}
\format{An object of class \code{data.frame} with 15 rows and 6 columns.}
\usage{
region_id_metadata
}
\description{
Simple toy metadata experimental conditions for 15 individuals in three experiments.
the condition (\code{test}) is linked to the experiments, and another condition (\code{treatment}) is per individual.
It serves as a simple example.
#'
}
\author{
Quentin Geissmann
}
\keyword{datasets}
