% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{scopr_example}
\alias{scopr_example}
\alias{scopr_example_dir}
\title{Get path to scopr example}
\usage{
scopr_example(path = NULL)

scopr_example_dir()
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
\code{scopr} comes with a sample DAM2 files in its \code{inst/extdata}
directory. \code{scopr_example} allow  make them easy to access.
}
\examples{
# list all files
scopr_example()
}
