% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{scoprConfiguration}
\alias{scoprConfiguration}
\title{Configuration of scopr}
\description{
Load and update the configuration of the program
}
\details{
Anywhere in scopr \code{scoprConfiguration} can be instantiated so
default values for hardware defined parameters can be set dynamically
via a config file
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{content}}{List of configuration items}

\item{\code{config_file}}{Default configuration file path
Initialize a configuration instance
Called by running \code{scoprConfiguration$new()}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-scoprConfiguration-new}{\code{scoprConfiguration$new()}}
\item \href{#method-scoprConfiguration-toggle}{\code{scoprConfiguration$toggle()}}
\item \href{#method-scoprConfiguration-verify}{\code{scoprConfiguration$verify()}}
\item \href{#method-scoprConfiguration-save}{\code{scoprConfiguration$save()}}
\item \href{#method-scoprConfiguration-load}{\code{scoprConfiguration$load()}}
\item \href{#method-scoprConfiguration-clone}{\code{scoprConfiguration$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scoprConfiguration-new"></a>}}
\if{latex}{\out{\hypertarget{method-scoprConfiguration-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scoprConfiguration$new(config_file = config_files)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Path to the default configuration file
Make sure all folders that scopr needs are writable
if they are not, recreate them as needed in the HOME folder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scoprConfiguration-toggle"></a>}}
\if{latex}{\out{\hypertarget{method-scoprConfiguration-toggle}{}}}
\subsection{Method \code{toggle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scoprConfiguration$toggle(property)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scoprConfiguration-verify"></a>}}
\if{latex}{\out{\hypertarget{method-scoprConfiguration-verify}{}}}
\subsection{Method \code{verify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scoprConfiguration$verify(config_file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Configuration file path
Save the configuration in to a config_file

Save the configuration stored in self$content
If the passed config_file is null, use the instance's default}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scoprConfiguration-save"></a>}}
\if{latex}{\out{\hypertarget{method-scoprConfiguration-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scoprConfiguration$save(config_file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Configuration file path
Load a a configuration from a config_file
If the passed config_file is null, use the instance's default
If the passed file does not exist create it with the conf in the default file
If it exists, load its contents and update the configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scoprConfiguration-load"></a>}}
\if{latex}{\out{\hypertarget{method-scoprConfiguration-load}{}}}
\subsection{Method \code{load()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scoprConfiguration$load(config_file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Configuration file path}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scoprConfiguration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-scoprConfiguration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scoprConfiguration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
