% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-ethoscope-metadata.R
\name{unlink_ethoscope_metadata}
\alias{unlink_ethoscope_metadata}
\title{Unlink an already linked metadata
This makes it possible to repeat an ethoscope data analysis from scratch}
\usage{
unlink_ethoscope_metadata(metadata)
}
\arguments{
\item{metadata}{A data.table containing a linked metadata
i.e. it should contain columns id, machine_id, datetime and file_info at least}
}
\value{
The same metadata after losing the linking information
}
\description{
Unlink an already linked metadata
This makes it possible to repeat an ethoscope data analysis from scratch
}
