% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_metadata.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Verify the pased metadata is compliant}
\usage{
validate_metadata(metadata)
}
\arguments{
\item{metadata}{A data.table describing an experiment
It must contain the columns region_id, machine_name and date_time
It can contain the column reference_hour to provide a difference ZT for groups of flies
It can contain any other column that is not used in the analysis
It must NOT contain columns with NA values, as that will make scopr ignore the row
date_time must follow the format \%YYYY-\%MM-\%DD_\%HH-\%MM-\%SS}
}
\description{
Verify the pased metadata is compliant
}
