% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-query.R
\name{parse_query}
\alias{parse_query}
\title{Build a query for loading ethoscope data}
\usage{
parse_query(x, result_dir = NULL, index_file = NULL)
}
\arguments{
\item{x}{object such as a \link{data.frame}, or the name of a file (see detail)}

\item{result_dir}{the root directory where all data are saved}

\item{index_file}{the name of an index_file, in \code{result_dir} (useful for loading remote data).}
}
\value{
a \link[data.table:data.table]{data.table::data.table}.
It contains the columns:
\itemize{
\item \code{path} -- the file to retreive the data from
\item \code{experiment_id} -- an autogenerated unique id for the experiment
\item \code{region_id} -- a unique identifier of each individual (region of interest) within an experiment
\item Variables already present in x (if x was a \code{data.frame}).
}
}
\description{
This function makes a query table, which is essential for subsequent loading of ethoscope data (via \link{query_ethoscopes}).
}
\details{
This function will generate a query from tree different types of inputs:
\enumerate{
\item A vector of names of \code{.db} files
\item A \link{data.frame} (recomended).
In this case, the function will try to match requested data with data available on \code{result_dir}.
The provided \link{data.table} has typically one row per requested individial and the columns
(not necessarily in this order):
\itemize{
\item \code{machine_name} -- the name of the machine in which the individual was (e.g. \code{"ETHOSCOPE_001"})
\item \code{date} -- the start date of the experiment formated as \code{"YYYY-MM-DD"}
\item \code{region_id} -- the ROI in which the animal was. When \emph{not provided, all regions are queried}.
\item \code{time} -- the start time of the experiment formated as "HH:MM:SS".
When \emph{not provided}, and multiple expriment for the same machine exist, \emph{only the last one is loaded}.
\item \code{???} -- any number of arbitrary columns* to associate \code{conditions}/\code{treatments}/\code{genotypes}/... to the previous columns.
}
}
}
\examples{
\donttest{
todo
}
}
\seealso{
\itemize{
\item \link{query_ethoscopes} -- to load the actual data
\item \link{list_result_files} -- to list available file
\item TODO -- to understand the magical power of the query system!
}
}
