% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-ethoscopes.R
\name{query_ethoscopes}
\alias{query_ethoscopes}
\title{Read data from ethoscope result files}
\usage{
query_ethoscopes(query, min_time = 0, max_time = Inf,
  reference_hour = NULL, verbose = TRUE, columns = NULL, ncores = 1,
  FUN = NULL, ...)
}
\arguments{
\item{query}{\link[data.table:data.table]{data.table::data.table} obtained by \link{parse_query}}

\item{min_time}{Exclude data before min_time (in seconds).
This time is \emph{relative to the start of the experiment}.}

\item{max_time}{Exclude data after max_time (in seconds).
It is also relative to the start of the experiment.}

\item{reference_hour}{Hour, in the day, to use as ZT0 reference.
When unspecified, time will be relative to the start of the experiment.}

\item{verbose}{whether to print progress (a logical)}

\item{columns}{optionnal vector of columns to be selected from the db file.
Time (t) is always implicitely selected.}

\item{ncores}{number of cores to use for optionnal parallel processing}

\item{FUN}{function (optional) to transform the data from each individual
immediately after is has been loaded.}

\item{...}{extra arguments to be passed to \code{FUN}}
}
\value{
a \link{behavr} table.
The metadata contains all the query columns and an autogenerated id per animal.
The data has the columns:
\itemize{
\item \code{id} -- autogenerated unique identifier, one per animal
\item \code{t} -- time
\item Several variables recorded by ethoscopes (position, angle, width/height and others).
Distance units (e.g. xy position, height/width) are expressed as a fraction of the width of the ROI they originate from.
}
}
\description{
This function is used to import behavioural data generated
by the \href{http://gilestrolab.github.io/ethoscope/}{ethoscope platform}.
That is it loads multiple \code{.db`` files into a single}R` \link[behavr:behavr]{behavr::behavr} table.
}
\details{
todo
}
\examples{

}
\seealso{
\itemize{
\item \link[behavr:behavr]{behavr::behavr} -- to understand the output format
\item \link{parse_query} -- to generate a query
\item \link{read_metadata} -- to show the metadata of a specific experiment
\item \link{list_result_files} -- to list available files
}
}
