% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Retrieve  metadata from an ethoscope result file}
\usage{
read_metadata(FILE)
}
\arguments{
\item{FILE}{the name of the input file}
}
\value{
A list containing fields for metadata entries
}
\description{
This function is used to obtain metadata -- such as \emph{time and date} of the experiment,
\emph{acquisition device}, \emph{version of the software} and such--
embedded in a result file (`.db``) generated by the ethoscope platform.
}
\seealso{
TODO
}
