% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Retrieve metadata from an ethoscope result file}
\usage{
read_metadata(FILE)
}
\arguments{
\item{FILE}{the name of the input file}
}
\value{
a list containing fields for metadata entries
}
\description{
This function is used to obtain metadata -- such as \emph{time and date} of the experiment,
\emph{acquisition device}, and \emph{version of the software} --
embedded in a result file (\code{.db}) generated by the ethoscope platform.
}
\seealso{
\itemize{
\item \link{query_ethoscopes} -- to load the actual data
\item \link{list_result_files} -- to list available files
}
}
